+++ src/libs/instance.ts
@@ -1,8 +1,8 @@
-import axios from 'axios';
+import axios from "axios";
 
 export const instance = axios.create({
   baseURL: process.env.NEXT_PUBLIC_API_URL,
   headers: {
-    'Content-Type': 'application/json',
-  },
-});
+    "Content-Type": "application/json",
+  }
+})
